+++ shared/src/Messaging/Enums.rs
