+++ shared/src/lib.rs
@@ -12,7 +12,7 @@ extern crate uuid;
 
 pub mod Constants;
 pub mod Locale;
-pub mod ProtocolMessages;
+pub mod Messaging;
 pub mod ProtocolTrait;
 pub mod TalkBack;
 pub mod Wallets;
